% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getConfirmedFormula}
\alias{getConfirmedFormula}
\alias{getNonRejectedFormula}
\title{Export Boruta result as a formula}
\usage{
getConfirmedFormula(x)

getNonRejectedFormula(x)
}
\arguments{
\item{x}{an object of a class Boruta, made using a formula interface.}
}
\value{
Formula, corresponding to the Boruta results.
\code{getConfirmedFormula} returns only Confirmed attributes, \code{getNonRejectedFormula} also adds Tentative ones.
}
\description{
Functions which convert the Boruta selection into a formula, so that it could be passed further to other functions.
}
\note{
This operation is possible only when Boruta selection was invoked using a formula interface.
}

