% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{getImpCV}
\alias{getImpCV}
\title{CV importance}
\usage{
getImpCV(x, y, nfolds = 5, getUnderlyingImp = getImpRfZ, ...)
}
\arguments{
\item{x}{data frame of predictors including shadows. Will be either H2O Frame or a vector.}

\item{y}{response vector. Will be either H2O Frame or a vector.}

\item{nfolds}{Number of folds (random data subsets) to use.}

\item{getUnderlyingImp}{Underlying importance function to call for each cross-validation subset.}

\item{...}{other parameters passed to the underlying h2o getImp call.}
}
\description{
This function is intended to be given to a \code{getImp} argument of \code{\link{Boruta}} function to be called by the Boruta algorithm as an importance source.
The getImp is called repeatedly on cross-validation subsets of the data.
}
