% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{getImpFru}
\alias{getImpFru}
\alias{getImpFruZ}
\alias{getImFruRaw}
\alias{getImpFruRaw}
\title{Fru Random Forest importance adapters}
\usage{
getImpFruZ(
  x,
  y,
  ntree = 500,
  trees = ntree,
  num.threads = 0,
  threads = num.threads,
  ...
)

getImpFruRaw(
  x,
  y,
  ntree = 500,
  trees = ntree,
  num.threads = 0,
  threads = num.threads,
  ...
)
}
\arguments{
\item{x}{data frame of predictors including shadows.}

\item{y}{response vector.}

\item{ntree}{Number of trees in the forest; copied into \code{\link[fru]{fru}}'s native \code{trees}, put to retain transparent compatibility with randomForest.}

\item{trees}{Number of trees in the forest, as according to \code{\link[fru]{fru}}'s nomenclature. If not given, set to \code{ntree} value. If both are given, \code{trees} takes precedence.}

\item{num.threads}{Number of computing threads to use; copied into \code{\link[fru]{fru}}'s native \code{threads}, put to retain transparent compatibility with ranger.}

\item{threads}{Number of computing threads to use, as according to \code{\link[fru]{fru}}'s nomenclature.
If not given, set to \code{num.threads}.
If both are given, \code{threads} takes precedence.
Default value of zero means all available threads.}

\item{...}{parameters passed to the underlying \code{\link[fru]{fru}} call; they are relayed from \code{...} of \code{\link{Boruta}}.}
}
\description{
Those function is intended to be given to a \code{getImp} argument of \code{\link{Boruta}} function to be called by the Boruta algorithm as an importance source.
\code{getImpFruZ} generates default, normalized permutation importance while \code{getImpFruRaw} raw permutation importance.
Fru does not support Gini index importance.
}
\note{
In prior versions of Boruta, other implementations of Random Forest were used as a default; prior to 5.0.0, randomForest package was used, and the default adapter used back then is available as \code{\link{getImpLegacyRf}}.
Prior to 10.0, ranger package was used with \code{\link{getImpRfZ}}.
}
