% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{getImpH2O}
\alias{getImpH2O}
\title{H2O importance}
\usage{
getImpH2O(x, y, h2o_model_fn = h2o::h2o.randomForest,
  importance_type = c("relative_importance", "scaled_importance",
  "percentage"), model_id = NULL, ...)
}
\arguments{
\item{x}{data frame of predictors including shadows. Will be either H2O Frame or a vector.}

\item{y}{response vector. Will be either H2O Frame or a vector.}

\item{h2o_model_fn}{H2O model to use for importance. Note that some models, such as h2o.naiveBayes, cannot be used because 
h2o.varimp() does not work for those models.}

\item{importance_type}{Whether to use relative_importance, scaled_importance, or percentage from the \code{h2o.varimp} result.}

\item{model_id}{If provided, will pass this model_id to the underlying h2o call, so the model can be retrieved later. Otherwise, the h2o model is removed.}

\item{...}{other parameters passed to the underlying h2o call.}
}
\description{
This function is intended to be given to a \code{getImp} argument of \code{\link{Boruta}} function to be called by the Boruta algorithm as an importance source.
}
