% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{getImpNJMIb}
\alias{getImpNJMIb}
\title{Normalised joint mutual information importance with bootstrap}
\usage{
getImpNJMIb(x, y, depth = 5, ...)
}
\arguments{
\item{x}{data frame of predictors including shadows.}

\item{y}{response vector.}

\item{depth}{maximal interaction depth.}

\item{...}{additional parameters relayed from \code{...} of \code{\link{Boruta}}, ignored.}
}
\description{
This function if intended to be given to a \code{getImp} argument of \code{\link{Boruta}} function to be called by the Boruta algorithm as an importance source.
}
