% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trandapters.R
\name{pairsTransdapter}
\alias{pairsTransdapter}
\title{Attribute pair transdapter}
\usage{
pairsTransdapter(
  adapter = getImpRfZ,
  reduce = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{adapter}{importance adapter to transform.}

\item{reduce}{function to reduce scores of all relevant pairs back into an attribute score.}
}
\value{
transformed importance adapter which can be fed into \code{getImp} argument of the \code{\link{Boruta}} function.
}
\description{
Wraps the importance adapter to work on order relations between pairs of attributes,
}
\examples{
set.seed(777)
data(iris)
Boruta(Species~.,data=iris,getImp=pairsTransdapter())
}
