\name{print.Boruta}
\alias{print.Boruta}
\title{Print Boruta object}
\usage{
  \method{print}{Boruta} (x, ...)
}
\arguments{
  \item{x}{an object of a class Boruta.}

  \item{...}{additional arguments passed to
  \code{\link{print}}.}
}
\value{
  Invisible copy of \code{x}.
}
\description{
  print method for Boruta objects.
}
\author{
  Miron B. Kursa
}

