% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boruta.R
\name{ResumeBoruta}
\alias{ResumeBoruta}
\title{Resume feature selection from Boruta object.}
\usage{
ResumeBoruta(x, y, priorBoruta, maxAdditionalRuns = 10,
  numPriorRuns = nrow(priorBoruta$ImpHistory), pValue = 0.01,
  mcAdj = TRUE, doTrace = 0, holdHistory = TRUE,
  getImp = getImpRfZ, ...)
}
\arguments{
\item{x}{data frame of predictors.}

\item{y}{response vector; factor for classification, numeric vector for regression, \code{Surv} object for survival (supports depends on importance adapter capabilities).}

\item{priorBoruta}{Boruta object.}

\item{maxAdditionalRuns}{Maximum number of additional importance source runs.}

\item{numPriorRuns}{Number of runs previously attempted. Not needed if the priorBoruta object contains importance history.}

\item{pValue}{confidence level. Default value should be used.}

\item{mcAdj}{if set to \code{TRUE}, a multiple comparisons adjustment using the Bonferroni method will be applied. Default value should be used; older (1.x and 2.x) versions of Boruta were effectively using \code{FALSE}.}

\item{doTrace}{verbosity level. 0 means no tracing, 1 means reporting decision about each attribute as soon as it is justified, 2 means the same as 1, plus reporting each importance source run, 3 means the same as 2, plus reporting of hits assigned to yet undecided attributes.}

\item{holdHistory}{if set to \code{TRUE}, the full history of importance is stored and returned as the \code{ImpHistory} element of the result.
Can be used to decrease a memory footprint of Boruta in case this side data is not used, especially when the number of attributes is huge; yet it disables plotting of such made \code{Boruta} objects and the use of the \code{\link{TentativeRoughFix}} function.}

\item{getImp}{function used to obtain attribute importance.
The default is getImpRfZ, which runs random forest from the \code{ranger} package and gathers Z-scores of mean decrease accuracy measure.
It should return a numeric vector of a size identical to the number of columns of its first argument, containing importance measure of respective attributes.
Any order-preserving transformation of this measure will yield the same result.
It is assumed that more important attributes get higher importance. +-Inf are accepted, NaNs and NAs are treated as 0s, with a warning.}

\item{...}{additional parameters passed to \code{getImp}.}
}
\description{
See \code{\link{Boruta}} for details of the feature selection algorithm. 
Run Boruta and then pass in that object to resume feature selection, when
features remain tentative.
}
\examples{
set.seed(777)
#Add some nonsense attributes to iris dataset by shuffling original attributes
iris.extended<-data.frame(iris,apply(iris[,-5],2,sample))
names(iris.extended)[6:9]<-paste("Nonsense",1:4,sep="")
#Run Boruta on this data
Boruta.iris.extended <- Boruta(y = iris.extended$Species, x = iris.extended[,-which(colnames(iris.extended) == "Species")], doTrace=2)
#Nonsense attributes should be rejected, except 1
print(Boruta.iris.extended)
Boruta.iris.extended2 <- ResumeBoruta(y = iris.extended$Species, x = iris.extended[,-which(colnames(iris.extended) == "Species")], priorBoruta = Boruta.iris.extended, maxAdditionalRuns=1000, doTrace=2)
#Last nonsense attribute shoudl be confirmed important.
print(Boruta.iris.extended)
print(Boruta.iris.extended2)
}
