% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boruta.R
\docType{data}
\name{srx}
\alias{srx}
\title{Small redundant XOR data}
\format{
A data frame with 8 predictors, 4 relevant: A, B, AoB, AnB and nA, as well as 3 irrelevant N1, N2 and N3, and decision attribute Y.
}
\source{
\url{https://blog.mbq.me/relevance-and-redundancy/}
}
\usage{
data(srx)
}
\description{
A synthetic data set with 32 rows corresponding to all combinations of values of five logical features, A, B, N1, N2 and N3.
The decision Y is equal to A xor B, hence N1--N3 are irrelevant attributes.
The set also contains 3 additional features, A or B (AoB), A and B (AnB) and not A (nA), which provide a redundant, but still relevant way to reconstruct Y.
}
\details{
This is set is an easy way to demonstrate the difference between all relevant feature selection methods, which should select all features except N1--N3, and minimal optimal ones, which will probably ignore most of them.
}
\keyword{datasets}
