% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Boruta}
\alias{Boruta}
\alias{Boruta.default}
\alias{Boruta.formula}
\title{Important attribute search using Boruta algorithm}
\usage{
Boruta(x, ...)

\method{Boruta}{default}(x, y, pValue = 0.01, mcAdj = TRUE, maxRuns = 100,
  doTrace = 0, holdHistory = TRUE, getImp = getImpRfZ, ...)

\method{Boruta}{formula}(formula, data = .GlobalEnv, ...)
}
\arguments{
\item{x}{data frame of predictors.}

\item{y}{response vector; factor for classification, numeric vector for regression.}

\item{getImp}{function used to obtain attribute importance. The default is getImpRfZ, which runs random forest from \code{randomForest} package and gathers Z-scores of mean decrease accuracy measure. It should return a numeric vector of a size identical to the number of columns of its first argument, containing importance measure of respective attributes. Any order-preserving transformation of this measure will yield the same result. It is assumed that more important attributes get higher importance. +-Inf are accepted, NaNs and NAs are treated as 0s, with a warning.}

\item{pValue}{confidence level. Default value should be used.}

\item{mcAdj}{if set to \code{TRUE}, a multiple comparisons adjustment using the Bonferroni method will be applied. Default value should be used; older (1.x and 2.x) versions of Boruta were effectively using \code{FALSE}.}

\item{maxRuns}{maximal number of importance source runs. You may increase it to resolve attributes left Tentative.}

\item{holdHistory}{if set to \code{TRUE}, the full history of importance is stored and returned as the \code{ImpHistory} element of the result. Can be used to decrease a memory footprint of Boruta in case this side data is not used, especially when the number of attributes is huge; yet it disables plotting of such made \code{Boruta} objects and the use of the \code{\link{TentativeRoughFix}} function.}

\item{doTrace}{verbosity level. 0 means no tracing, 1 means printing a "." sign after each importance source run,
2 means same as 1, plus consecutive reporting of test results.}

\item{...}{additional parameters passed to \code{getImp}.}

\item{formula}{alternatively, formula describing model to be analysed.}

\item{data}{in which to interpret formula.}
}
\value{
An object of class \code{Boruta}, which is a list with the following components:
\item{finalDecision}{a factor of three value: \code{Confirmed}, \code{Rejected} or \code{Tentative},
containing final result of feature selection.}
\item{ImpHistory}{a data frame of importances of attributes gathered in each importance source run.
Beside predictors' importances contains maximal, mean and minimal importance of shadow attributes in each run.
Rejected attributes get \code{-Inf} importance. Set to \code{NULL} if \code{holdHistory} was given \code{FALSE}.}
\item{timeTaken}{time taken by the computation.}
\item{impSource}{string describing the source of importance, equal to a comment attribute of the \code{getImp} argument.}
\item{call}{the original call of the \code{Boruta} function.}
}
\description{
Boruta is an all-relevant feature selection wrapper algorithm.
It finds relevant features by comparing original attributes'
importance with importance achievable at random, estimated
using their permuted copies.
}
\details{
Boruta iteratively compares importances of attributes with importances of shadow attributes, created by
shuffling original ones. Attributes that have significantly worst importance than shadow ones
are being consecutively dropped. On the other hand, attributes that are significantly better than
shadows are admitted to be Confirmed. Shadows are re-created in each iteration.
Algorithm stops when only Confirmed attributes are left, or when it reaches \code{maxRuns} importance source runs
in the last round. If the second scenario occurs, some attributes may be left without a decision. They are
claimed Tentative. You may try to extend \code{maxRuns} or lower \code{pValue} to clarify them, but in
some cases their importances do fluctuate too much for Boruta to converge.
Instead, you can use \code{\link{TentativeRoughFix}} function, which will perform other, weaker test to make a final
decision, or simply treat them as undecided in further analysis.
}
\note{
Version 2.0.0 changes some name conventions and thus may be incompatible with scripts written for 1.x.x version and
old Boruta objects. Solutions of most problems of this kind should boil down to change of \code{ZScoreHistory} to \code{ImpHistory}
in script source or Boruta object structure.
}
\examples{
set.seed(777);
#Add some nonsense attributes to iris dataset by shuffling original attributes
iris.extended<-data.frame(iris,apply(iris[,-5],2,sample));
names(iris.extended)[6:9]<-paste("Nonsense",1:4,sep="");
#Run Boruta on this data
Boruta(Species~.,data=iris.extended,doTrace=2)->Boruta.iris.extended
#Nonsense attributes should be rejected
print(Boruta.iris.extended);
#Boruta using rFerns' importance
Boruta(Species~.,data=iris.extended,getImp=getImpFerns)->Boruta.ferns.irisE
print(Boruta.ferns.irisE);
\dontrun{
#Boruta on the Ozone data from mlbench
library(mlbench); data(Ozone);
na.omit(Ozone)->ozo;
#Takes some time, so be patient
Boruta(V4~.,data=ozo,doTrace=2)->Bor.ozo;
cat('Random forest run on all attributes:\\n');
print(randomForest(V4~.,data=ozo));
cat('Random forest run only on confirmed attributes:\\n');
print(randomForest(ozo[,getSelectedAttributes(Bor.ozo)],ozo$V4));
}
\dontrun{
#Boruta on the HouseVotes84 data from mlbench
library(mlbench); data(HouseVotes84);
na.omit(HouseVotes84)->hvo;
#Takes some time, so be patient
Boruta(Class~.,data=hvo,doTrace=2)->Bor.hvo;
print(Bor.hvo);
plot(Bor.hvo);
}
\dontrun{
#Boruta on the Sonar data from mlbench
library(mlbench); data(Sonar);
#Takes some time, so be patient
Boruta(Class~.,data=Sonar,doTrace=2)->Bor.son;
print(Bor.son);
#Shows important bands
plot(Bor.son,sort=FALSE);
}
}
\author{
Miron B. Kursa, based on the idea & original code by Witold R. Rudnicki.
}
\references{
Miron B. Kursa, Witold R. Rudnicki (2010). Feature Selection with the Boruta Package.
\emph{Journal of Statistical Software, 36(11)}, p. 1-13.
URL: \url{http://www.jstatsoft.org/v36/i11/}
}

