% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trandapters.R
\name{conditionalTransdapter}
\alias{conditionalTransdapter}
\title{Conditional transdapter}
\usage{
conditionalTransdapter(groups, adapter = getImpRfZ)
}
\arguments{
\item{groups}{groups.}

\item{adapter}{importance adapter to transform.}
}
\value{
transformed importance adapter which can be fed into \code{getImp} argument of the \code{\link{Boruta}} function.
}
\description{
Applies downstream importance source on a given object strata and averages their outputs.
}
