% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSelectedAttributes}
\alias{getSelectedAttributes}
\title{Extract names of the selected attributes}
\usage{
getSelectedAttributes(x, withTentative = FALSE)
}
\arguments{
\item{x}{an object of a class Boruta, from which relevant attributes names should be extracted.}

\item{withTentative}{if set to \code{TRUE}, Tentative attributes will be also returned.}
}
\value{
A character vector with names of the relevant attributes.
}
\description{
Extract names of the selected attributes
}
\examples{
\dontrun{
data(iris);
#Takes some time, so be patient
Boruta(Species~.,data=iris,doTrace=2)->Bor.iris;
print(Bor.iris);
print(getSelectedAttributes(Bor.iris));
}
}
\author{
Miron B. Kursa
}

