% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boruta.R
\name{h2o.shuffle_column}
\alias{h2o.shuffle_column}
\title{Shuffle H2O column.}
\usage{
h2o.shuffle_column(dat, col_idx)
}
\arguments{
\item{dat}{An H2O frame or something that can be converted to one using as.h2o.}

\item{col_idx}{Column index for the column to be shuffled.}
}
\value{
Shuffled H2O frame.
}
\description{
Helper function to shuffle a single H2o column.
}
