% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boruta.R
\name{print.Boruta}
\alias{print.Boruta}
\title{Print Boruta object}
\usage{
\method{print}{Boruta}(x, ...)
}
\arguments{
\item{x}{an object of a class Boruta.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible copy of \code{x}.
}
\description{
Print method for the Boruta objects.
}
\author{
Miron B. Kursa
}

